project/requirements.txt
weaviate-client[agents]>=4.1.1
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
pydantic>=2.5.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

